Add solenoid valve and Zolix Omni-λ device implementations
- Introduced solenoid_valve_4v110.yaml for the 4V110 solenoid valve configuration with action mappings and status types.
- Updated xyz_guangdian.py to change error_code return type from int to float for consistency.
- Corrected module path in xyz_guangdian.yaml to reflect the correct import structure.
- Added graph_example_zolix_omni_lambda.json for Zolix Omni-λ device representation.
- Implemented zolix_omni_lambda.py with comprehensive device control and communication methods.
- Created zolix_omni_lambda.yaml for Zolix Omni-λ device configuration, including action mappings and status types.
- Updated pyproject.toml and requirements.txt to include new dependencies for device communication and processing.

+++ device_package_example/devices/xyz_guangdian/xyz_guangdian.py
@@ -314,9 +314,9 @@ class XYZGuangdian:
         return self.data.get("push_rod_status", "released")
 
     @property
-    def error_code(self) -> int:
+    def error_code(self) -> float:
         """错误代码"""
-        return self.data.get("error_code", 0)
+        return float(self.data.get("error_code", 0))
 
     # ========== 动作方法 ==========
